requirements.txt
Django==4.2
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
setuptools
python-dotenv
grpcio
grpcio-tools
protobuf==4.22.3
psycopg2-binary==2.9.6
PyJWT==2.6.0
redis==4.5.4
drf-yasg
unidecode